Basic Inventory System Setup
How does one say "Hardcoded" in French? This is basic baseline of inventory system. This method works but is really not the best way to do this, I just have become very frustrated and my eyes hurt from staring at the screen. Right now, allows player to place the minion they want from picking it up in the game into their combat slot. This is almost all UI right now and has not been implemented to work with combat yet.